% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{FileList}
\alias{FileList}
\title{Get list of avialable files}
\usage{
FileList(FileMask = "*")
}
\arguments{
\item{FileMask}{- File mask for file search Applied to file name only without extension. Case insensitive. Can contain "*"and "?" mask symbols}
}
\value{
List of available file names
}
\description{
Get list of avialable files by provided mask
Corresponds to mzAccess web-service API function FileList
}
\examples{
FileList("*72h*")
}
