% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetArea}
\alias{GetArea}
\title{Get LC-MS Area signals}
\usage{
GetArea(FileName, MZLow, MZHigh, RTLow, RTHigh, Cache = TRUE,
  Profile = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{MZLow}{- Minimum m/z value for LC-MS area requested}

\item{MZHigh}{- Maximum m/z value for LC-MS area requested}

\item{RTLow}{- Minimum retention time value for LC-MS area requested}

\item{RTHigh}{- Maximum retention time value for LC-MS area requested}

\item{Cache}{- If TRUE data will be loaded from fast access cache, if FALSE - from original raw files}

\item{Profile}{- If TRUE data will presented in profile mode how is was acquired by mass spectrometer, If FALSE data will be presented in peak centroided mode}
}
\value{
Data frame of Retention Time and Intensities for requested LC-MS area
}
\description{
Gets all non-zero signals for specified LC-MS area.
Corresponds to mzAccess web-service API function GetArea
}
\examples{
GetChromatogram("031_MOO_Labeling_HIL_72h_100_3d_MM",148.0584342,148.0624342,0,16)
}

