% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetChromatogramArray}
\alias{GetChromatogramArray}
\title{Get Array of XIC Chromatogram}
\usage{
GetChromatogramArray(FileNames, MZLows, MZHighs, RTLows, RTHighs,
  Cache = TRUE)
}
\arguments{
\item{FileNames}{- List of names of original raw mass spectrometry file.}

\item{Cache}{- If TRUE data will be loaded from fast access cache, if FALSE - from original raw files}

\item{MZLow}{- List of minimum m/z values for LC-MS areas requested}

\item{MZHigh}{- List of maximum m/z values for LC-MS areas requested}

\item{RTLow}{- List of minimum retention time values for LC-MS areas requested}

\item{RTHigh}{- List of maximum retention time values for LC-MS areas requested}
}
\value{
List of data frames. Each n-th data frame represent single XIC chromatogram specified
 by n-th members of incoming arrays and consist of
 of Retention Time and Intensities for requested LC-MS area
}
\description{
Batch analog of GetChromatogram function.
Gets array of eXtracted Ion Current (XIC) chromatograms for specified LC-MS areas.
Equal length of incoming lists is assumed.
Corresponds to mzAccess web-service API function GetChromatogramsArray
}
\examples{

}

