% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetFragmentationEvents}
\alias{GetFragmentationEvents}
\title{Get data frame of fragmentation events in requested LC-MS area}
\usage{
GetFragmentationEvents(FileName, MZLow, MZHigh, RTLow, RTHigh)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file.}

\item{MZLow}{- Minimum parent ion m/z value to be included in event list}

\item{MZHigh}{- Maximum parent ion m/z value to be included in event list}

\item{RTLow}{- Minimum retention time value to be included in event list}

\item{RTHigh}{- Maximum retention time value to be included in event list}
}
\value{
Data frame of fragmentation events inside of desired LC-MS region
Includes Scan Number (to use with function GetSpectraByScanNumber),
Mass of parent ion,
Retention time of fragmentation event
MSOrder - if 2 - that is MSMS spectra, if more it is high order MSN spectra
Desc - description of spectrum
}
\description{
Get data frame of fragmentation events in requested LC-MS area
Corresponds to mzAccess web-service API function GetFragmentationEvents
}
\examples{
GetFragmentationEvents("Agilent_QTOF_plasma_1.d",200,250,6,6.5)
}

