% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetMZRange}
\alias{GetMZRange}
\title{Get mass range for spectra in file}
\usage{
GetMZRange(FileName, Cache = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{Cache}{- If TRUE Mass Range will show minimum amd maximim masses avialable in spectra through the file, if FALSE - it will return Mass Interval where masses were scanned by mass spectrometer}
}
\value{
MassRange object with slots MinMZ,MaxMZ
}
\description{
Get mass range for spectra in file
Corresponds to mzAccess web-service API function GetMassRange
}
\examples{
GetMZRange("Thermo_QE_cells_72h_LA_3")
}

