% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetRTRange}
\alias{GetRTRange}
\title{Get retention time range for spectra in file}
\usage{
GetRTRange(FileName, Cache = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{Cache}{- If TRUE RT Range will show RTs for first and last spectra in spectra through the file, if FALSE - it will retention time range as registered by mass spectrometer}
}
\value{
RTRange object with slots MinRT,MaxRT
}
\description{
Get retention time range for spectra in file
Corresponds to mzAccess web-service API function GetRTRange
}
\examples{
GetRTRange("031_MOO_Labeling_HIL_72h_100_3d_MM")
}

