% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetScanNumberFromRT}
\alias{GetScanNumberFromRT}
\title{Get scan number for retention time}
\usage{
GetScanNumberFromRT(FileName, RT, Cache = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{RT}{- retention time for which spectrum will be searched.}

\item{Cache}{- If TRUE scan number will be loaded from fast access cache, if FALSE - from original raw files}
}
\value{
Interger scan number
}
\description{
Get scan number
Corresponds to mzAccess web-service API function GetScanNumberFromRT
Web-service will provide first scan number  where RT is not less than specified
}
\examples{
GetScanNumberFromRT("031_MOO_Labeling_HIL_72h_100_3d_MM",6)
}

