% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetSpectrumByRT}
\alias{GetSpectrumByRT}
\title{Get spectrum for  RT}
\usage{
GetSpectrumByRT(FileName, MZLow, MZHigh, RT, Cache = FALSE, Profile = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{MZLow}{- Minimum m/z value to be included in spectrum}

\item{MZHigh}{- Maximum m/z value to be included in spectrum}

\item{RT}{- retention time for which spectrum will be searched.}

\item{Cache}{- If TRUE data will be loaded from fast access cache, if FALSE - from original raw files}

\item{Profile}{- If TRUE data will presented in profile mode how is was acquired by mass spectrometer, If FALSE data will be presented in peak centroided mode}
}
\value{
Data frame of Mass and Intensities for requested LC-MS area
}
\description{
Get Single scan spectrum on certain RT.
Web-service will provide first spectrum where RT is not less than specified
Corresponds to mzAccess web-service API function GetSpectrumbyRT
}
\examples{
GetSpectrumByRT("Thermo_QE_cells_72h_LA_3", MZLow = 50, MZHigh = 400, RT = 6, Cache = FALSE, Profile = FALSE)
}
