% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetSpectrumbyScanNumber}
\alias{GetSpectrumbyScanNumber}
\title{Get spectrum for scan number}
\usage{
GetSpectrumbyScanNumber(FileName, MZLow, MZHigh, ScanNumber, Cache = TRUE,
  Profile = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{MZLow}{- Minimum m/z value to be included in spectrum}

\item{MZHigh}{- Maximum m/z value to be included in spectrum}

\item{ScanNumber}{- scan number for requested spectrum.}

\item{Cache}{- If TRUE data will be loaded from fast access cache, if FALSE - from original raw files}

\item{Profile}{- If TRUE data will presented in profile mode how is was acquired by mass spectrometer, If FALSE data will be presented in peak centroided mode}
}
\value{
Data frame of Mass and Intensities for requested LC-MS area
}
\description{
Get Single scan spectrum for certain scan.
Corresponds to mzAccess web-service API function GetSpectrumbyScanNumber
}
\examples{
GetChromatogram("031_MOO_Labeling_HIL_72h_100_3d_MM",148.0584342,148.0624342,0,16)
}

