% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetRTFromScanNumber}
\alias{GetRTFromScanNumber}
\title{Get retention time for scan number}
\usage{
GetRTFromScanNumber(FileName, ScanNumber, Cache = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{Cache}{- If TRUE retention time value (in minutes) will be loaded from fast access cache, if FALSE - from original raw files}

\item{RT}{- retention time for which spectrum will be searched.}
}
\value{
Retention time value
}
\description{
Get retention time for scan number
Corresponds to mzAccess web-service API function GetRTFromScanNumber
Web-service will provide first scan number  where RT is not less than specified
}
\examples{
GetRTFromScanNumber("Thermo_QE_cells_72h_LA_3",1000)
}
