% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetSpectrumArray}
\alias{GetSpectrumArray}
\title{Get Array of averaged spectra}
\usage{
GetSpectrumArray(FileNames, MZLows, MZHighs, RTLows, RTHighs, Profile = FALSE)
}
\arguments{
\item{Profile}{- If TRUE data will presented in profile mode how is was acquired by mass spectrometer, If FALSE data will be presented in peak centroided mode}

\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{MZLow}{- List of minimum m/z values to be included in spectrum}

\item{MZHigh}{- List of maximum m/z values to be included in spectrum}

\item{RTLow}{- List of minimum retention time values for RT range of averaged spectrum}

\item{RTHigh}{- List of maximum retention time values for RT range of averaged spectrum}
}
\value{
List of data frames of Mass and Intensities for requested averaged spectra
}
\description{
Batch analog of GetAveSpectrum  function.
Get Array of spectra averaged (or summed) within desired RT intervals
Corresponds to mzAccess web-service API function GetSpectrumArray
Equal length of incoming lists is assumed.
Averaging algorithm is  vendor specific, therefore function has no cache option
}
\examples{
GetAvgSpectrum("031_MOO_Labeling_HIL_72h_100_3d_MM", MZLow = 50, MZHigh = 400, RTLow = 6, RTHigh = 8, Profile = FALSE)
}

