% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSInterface.R
\name{GetSpectrumByScanNumber}
\alias{GetSpectrumByScanNumber}
\title{Get spectrum for scan number}
\usage{
GetSpectrumByScanNumber(FileName, MZLow, MZHigh, ScanNumber, Cache = FALSE,
  Profile = FALSE)
}
\arguments{
\item{FileName}{- Name of original raw mass spectrometry file. Can be stated with or without path and extention}

\item{MZLow}{- Minimum m/z value to be included in spectrum}

\item{MZHigh}{- Maximum m/z value to be included in spectrum}

\item{ScanNumber}{- scan number for requested spectrum.}

\item{Cache}{- If TRUE data will be loaded from fast access cache, if FALSE - from original raw files}

\item{Profile}{- If TRUE data will presented in profile mode how is was acquired by mass spectrometer, If FALSE data will be presented in peak centroided mode}
}
\value{
Data frame of Mass and Intensities for requested LC-MS area
}
\description{
Get Single scan spectrum for certain scan.
Corresponds to mzAccess web-service API function GetSpectrumbyScanNumber
}
\examples{
GetSpectrumByScanNumber("Thermo_QE_cells_72h_LA_3", MZLow = 50, MZHigh = 400, ScanNumber = 1000, Profile = TRUE)
}

